+++ react-for-beginners/src/App.js
@@ -1,30 +1,13 @@
 import { useState, useEffect } from "react";
+import Button from "./Button";
+import TodoList from "./TodoList";
 
 function App() {
-  const [toDo, setToDo] = useState("");
-  const [toDos, setToDos] = useState([]);
-  const onChange = (event) => setToDo(event.target.value);
-  // console.log(toDo);
-  const onSubmit = (event) => {
-    event.preventDefault();
-    if(toDo === "") {
-      return;
-    }
-    setToDos((currentArray) => [toDo, ...currentArray]);
-    setToDo("");
-  };
-  console.log(toDos);
+  
   return (
     <div>
-      <h1>My To Dos ({ toDos.length })</h1>
-      <form onSubmit={ onSubmit }>
-        <input
-          onChange={ onChange }
-          value={ toDo }
-          type="text"
-          placeholder="Write your to do...." />
-        <button>Add To Do</button>
-      </form>
+      <Button text="test button"/>
+      <TodoList />
     </div>
   );
 }
